+++ src/any_vec_ptr.rs
@@ -1,28 +1,46 @@
 //! Type dispatched analog of `enum{*AnyVecRaw, *AnyVec<Traits>}`.
 
+use std::marker::PhantomData;
 use std::ptr::NonNull;
+use crate::any_value::Unknown;
 use crate::any_vec_raw::AnyVecRaw;
 use crate::AnyVec;
 use crate::traits::Trait;
 
 pub trait IAnyVecRawPtr: Copy{
+    /// Known element type of AnyVec
+    type Element: 'static/* = Unknown*/;
     fn any_vec_raw(&self) -> NonNull<AnyVecRaw>;
 }
 pub trait IAnyVecPtr<Traits: ?Sized + Trait>: IAnyVecRawPtr{
     fn any_vec(&self) -> NonNull<AnyVec<Traits>>;
 }
 
-#[derive(Copy, Clone)]
-pub struct AnyVecRawPtr{
-    ptr: NonNull<AnyVecRaw>
+
+pub struct AnyVecRawPtr<Type: 'static>{
+    ptr: NonNull<AnyVecRaw>,
+    phantom: PhantomData<*mut Type>
 }
-impl From<NonNull<AnyVecRaw>> for AnyVecRawPtr {
+impl<Type> From<NonNull<AnyVecRaw>> for AnyVecRawPtr<Type>{
     #[inline]
     fn from(ptr: NonNull<AnyVecRaw>) -> Self {
-        Self{ptr}
+        Self{ptr, phantom: PhantomData}
     }
 }
-impl IAnyVecRawPtr for AnyVecRawPtr{
+impl<Type> Copy for AnyVecRawPtr<Type> {}
+impl<Type> Clone for AnyVecRawPtr<Type> {
+    #[inline]
+    fn clone(&self) -> Self {
+        Self{
+            ptr: self.ptr,
+            phantom: PhantomData
+        }
+    }
+}
+
+impl<Type> IAnyVecRawPtr for AnyVecRawPtr<Type>{
+    type Element = Type;
+
     #[inline]
     fn any_vec_raw(&self) -> NonNull<AnyVecRaw> {
         self.ptr
@@ -45,6 +63,12 @@ impl<Traits: ?Sized + Trait> From<&mut AnyVec<Traits>> for AnyVecPtr<Traits> {
         Self{ptr: NonNull::from(reference)}
     }
 }
+impl<Traits: ?Sized + Trait> From<&AnyVec<Traits>> for AnyVecPtr<Traits> {
+    #[inline]
+    fn from(reference: &AnyVec<Traits>) -> Self {
+        Self{ptr: NonNull::from(reference)}
+    }
+}
 impl<Traits: ?Sized + Trait> Clone for AnyVecPtr<Traits>{
     #[inline]
     fn clone(&self) -> Self {
@@ -54,6 +78,8 @@ impl<Traits: ?Sized + Trait> Clone for AnyVecPtr<Traits>{
 impl<Traits: ?Sized + Trait> Copy for AnyVecPtr<Traits>{}
 
 impl<Traits: ?Sized + Trait> IAnyVecRawPtr for AnyVecPtr<Traits> {
+    type Element = Unknown;
+
     #[inline]
     fn any_vec_raw(&self) -> NonNull<AnyVecRaw> {
         NonNull::from(unsafe{&self.ptr.as_ref().raw})