+++ src/any_vec_typed.rs
@@ -1,9 +1,13 @@
 use std::marker::PhantomData;
+use std::ops::{Range, RangeBounds};
 use std::ptr::NonNull;
+use std::slice;
 use crate::any_value::{AnyValue, AnyValueWrapper};
 use crate::any_vec_raw::AnyVecRaw;
 use crate::ops::{remove, swap_remove, TempValue};
-use crate::ops::any_vec_ptr::AnyVecRawPtr;
+use crate::any_vec_ptr::AnyVecRawPtr;
+use crate::into_range;
+use crate::ops::drain::Drain;
 
 /// Concrete type [`AnyVec`] representation.
 ///
@@ -60,8 +64,8 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
     pub fn remove(&mut self, index: usize) -> T {
         self.this().index_check(index);
         unsafe{
-            TempValue::<_>::new(remove::Remove::<_, T>::new(
-                AnyVecRawPtr::from(self.any_vec),
+            TempValue::<_>::new(remove::Remove::new(
+                AnyVecRawPtr::<T>::from(self.any_vec),
                 index
             )).downcast_unchecked::<T>()
         }
@@ -71,11 +75,22 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
     pub fn swap_remove(&mut self, index: usize) -> T {
         self.this().index_check(index);
         unsafe{
-            TempValue::<_>::new(swap_remove::SwapRemove::<_, T>::new(
-                AnyVecRawPtr::from(self.any_vec),
+            TempValue::<_>::new(swap_remove::SwapRemove::new(
+                AnyVecRawPtr::<T>::from(self.any_vec),
                 index
             )).downcast_unchecked::<T>()
         }
+    }
+
+    pub fn drain(&mut self, range: impl RangeBounds<usize>) -> impl Iterator<Item = T> {
+        let Range{start, end} = into_range(self.len(), range);
+        Drain::new(
+            AnyVecRawPtr::<T>::from(self.any_vec),
+            start,
+            end
+        ).map(|e| unsafe{
+            e.downcast_unchecked::<T>()
+        })
     }
 
     #[inline]
@@ -84,9 +99,39 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
     }
 
     #[inline]
+    pub fn iter(&self) -> slice::Iter<'a, T> {
+        self.as_slice().iter()
+    }
+
+    #[inline]
+    pub fn iter_mut(&mut self) -> slice::IterMut<'a, T> {
+        self.as_mut_slice().iter_mut()
+    }
+
+    #[inline]
+    pub fn get(&self, index: usize) -> Option<&'a T> {
+        self.as_slice().get(index)
+    }
+
+    #[inline]
+    pub unsafe fn get_unchecked(&self, index: usize) -> &'a T {
+        self.as_slice().get_unchecked(index)
+    }
+
+    #[inline]
+    pub fn get_mut(&mut self, index: usize) -> Option<&'a mut T>{
+        self.as_mut_slice().get_mut(index)
+    }
+
+    #[inline]
+    pub unsafe fn get_unchecked_mut(&mut self, index: usize) -> &'a mut T {
+        self.as_mut_slice().get_unchecked_mut(index)
+    }
+
+    #[inline]
     pub fn as_slice(&self) -> &'a [T] {
         unsafe{
-            std::slice::from_raw_parts(
+            slice::from_raw_parts(
                 self.this().mem.as_ptr().cast::<T>(),
                 self.this().len,
             )
@@ -96,7 +141,7 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
     #[inline]
     pub fn as_mut_slice(&mut self) -> &'a mut[T] {
         unsafe{
-            std::slice::from_raw_parts_mut(
+            slice::from_raw_parts_mut(
                 self.this_mut().mem.as_ptr().cast::<T>(),
                 self.this().len,
             )