+++ src/ops/mod.rs
@@ -1,11 +1,11 @@
 mod temp;
-pub(crate) mod any_vec_ptr;
 pub(crate) mod swap_remove;
 pub(crate) mod remove;
+pub(crate) mod drain;
 
 pub use temp::TempValue;
 
-use any_vec_ptr::AnyVecPtr;
+use crate::any_vec_ptr::AnyVecPtr;
 
 /// Lazily `remove` element on consumption/drop.
 ///
@@ -20,3 +20,10 @@ pub type Remove<'a, Traits> = TempValue<remove::Remove<'a, AnyVecPtr<Traits>>, T
 ///
 /// [`AnyVec::swap_remove`]: crate::AnyVec::swap_remove
 pub type SwapRemove<'a, Traits> = TempValue<swap_remove::SwapRemove<'a, AnyVecPtr<Traits>>, Traits>;
+
+/// Same as [`std::vec::Drain`].
+///
+/// This is created by [`AnyVec::drain`].
+///
+/// [`AnyVec::drain`]: crate::AnyVec::drain
+pub type Drain<'a, Traits> = drain::Drain<'a, AnyVecPtr<Traits>>;