+++ tests/any_vec_typed.rs
@@ -71,6 +71,53 @@ fn swap_remove_test() {
 }
 
 #[test]
+fn drain_all_test() {
+    let mut any_vec: AnyVec = AnyVec::new::<String>();
+    let mut vec = any_vec.downcast_mut::<String>().unwrap();
+    vec.push(String::from("0"));
+    vec.push(String::from("1"));
+    vec.push(String::from("2"));
+    vec.push(String::from("3"));
+
+    let mut vec2 = Vec::new();
+    for e in vec.drain(..){
+        vec2.push(e);
+    }
+    assert_eq!(any_vec.len(), 0);
+    assert_equal(vec2, [
+        String::from("0"),
+        String::from("1"),
+        String::from("2"),
+        String::from("3"),
+    ]);
+}
+
+#[test]
+fn any_vec_drain_in_the_middle_test() {
+    let mut any_vec: AnyVec = AnyVec::new::<String>();
+    let mut vec = any_vec.downcast_mut::<String>().unwrap();
+    vec.push(String::from("0"));
+    vec.push(String::from("1"));
+    vec.push(String::from("2"));
+    vec.push(String::from("3"));
+    vec.push(String::from("4"));
+
+    let mut vec2 = Vec::new();
+    for e in vec.drain(1..3){
+        vec2.push(e);
+    }
+    assert_equal(vec2, [
+        String::from("1"),
+        String::from("2"),
+    ]);
+    assert_equal(any_vec.downcast_ref::<String>().unwrap().as_slice(), &[
+        String::from("0"),
+        String::from("3"),
+        String::from("4"),
+    ]);
+}
+
+#[test]
 pub fn downcast_mut_test(){
     let mut any_vec: AnyVec = AnyVec::new::<String>();
     let mut vec = any_vec.downcast_mut::<String>().unwrap();
@@ -86,4 +133,19 @@ pub fn downcast_ref_test(){
     let vec2 = any_vec.downcast_ref::<String>().unwrap();
     assert_eq!(vec1.len(), 0);
     assert_eq!(vec2.len(), 0);
+}
+
+#[test]
+fn any_vec_into_iter_test() {
+    let mut any_vec: AnyVec = AnyVec::new::<usize>();
+    let mut vec = any_vec.downcast_mut::<usize>().unwrap();
+    vec.push(1);
+    vec.push(10);
+    vec.push(100);
+
+    let mut sum = 0;
+    for e in vec{
+        sum += *e;
+    }
+    assert_eq!(sum, 111);
 }